+++ src/problem_2.rs
@@ -1,5 +1,5 @@
-pub fn problem_2() -> Result<u32, String> {
-    const MAX_VAL: u32 = 4_000_000;
+pub fn problem_2() -> Result<u64, String> {
+    const MAX_VAL: u64 = 4_000_000;
     let mut n1 = 1;
     let mut n2 = 2;
     let mut sum = n2;
